Fix specifications for parameter type
