+++ src/core/node/nodeParameter.ts
@@ -26,14 +26,6 @@ export class NodeParameter extends Parameter {
   }
 
   /**
-   * Reset the constant value taken from the node component.
-   */
-  override reset(): void {
-    this.type = 'constant';
-    this.value = this.options.value;
-  }
-
-  /**
    * Trigger changes when the parameter is changed.
    */
   override paramChanges(): void {
@@ -56,14 +48,14 @@ export class NodeParameter extends Parameter {
       param.factors = this.factors;
     }
 
-    // Add the parameter type if not constant.
-    if (!this.isConstant) {
-      param.type = this.type;
-    }
-
     // Add the rules for validation if existed.
     if (this.rules.length > 0) {
       param.rules = this.rules;
+    }
+
+    // Add param type if not constant.
+    if (!this.isConstant) {
+      param.type = this.typeToJSON()
     }
 
     return param;
